+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- If one applies an empty schema using the 'ddl-manager' role API,
+  put in the clusterwide configuration an example instead.
+
 ## [1.4.0] - 2021-04-22
 
 ### Added